# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CreateInstitutionBody": {
        "properties": {
          "kind": {
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "kind",
          "name"
        ],
        "type": "object"
      },
...
      "CreateTypeBody": {
        "properties": {
          "categoryCode": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldDef"
            },
            "type": "array"
          },
          "icon": {
            "type": "string"
          },
          "label": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "fields",
          "label"
        ],
        "type": "object"
      },
...
      "Dashboard": {
        "properties": {
          "attention": {
            "items": {
              "$ref": "#/components/schemas/AttentionItem"
            },
            "type": "array"
          },
          "byCategory": {
            "items": {
              "$ref": "#/components/schemas/Breakdown"
            },
            "type": "array"
          },
          "byInstitution": {
            "items": {
              "$ref": "#/components/schemas/Breakdown"
            },
            "type": "array"
          },
          "byLiabilityKind": {
            "items": {
              "$ref": "#/components/schemas/Breakdown"
            },
            "type": "array"
          },
          "byMember": {
            "items": {
              "$ref": "#/components/schemas/Breakdown"
            },
            "type": "array"
          },
          "currency": {
            "type": "string"
          },
          "disclaimer": {
            "type": "string"
          },
          "holdingCount": {
            "format": "int32",
            "type": "integer"
          },
          "liabilityCount": {
            "format": "int32",
            "type": "integer"
          },
          "netWorth": {
            "type": "number"
          },
          "netWorthFormatted": {
            "type": "string"
          },
          "netWorthInWords": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "scopeLabel": {
            "type": "string"
          },
          "totalAssets": {
            "type": "number"
          },
          "totalAssetsFormatted": {
            "type": "string"
          },
          "totalLiabilities": {
            "type": "number"
          },
          "totalLiabilitiesFormatted": {
            "type": "string"
          },
          "upcoming": {
            "items": {
              "$ref": "#/components/schemas/UpcomingItem"
            },
            "type": "array"
          },
          "valueConfidence": {
            "$ref": "#/components/schemas/ValueConfidence"
          }
        },
        "required": [
          "attention",
          "byCategory",
          "byInstitution",
          "byLiabilityKind",
          "byMember",
          "currency",
          "disclaimer",
          "holdingCount",
          "liabilityCount",
          "netWorth",
          "netWorthFormatted",
          "netWorthInWords",
          "scope",
          "scopeLabel",
          "totalAssets",
          "totalAssetsFormatted",
          "totalLiabilities",
          "totalLiabilitiesFormatted",
          "upcoming",
          "valueConfidence"
        ],
        "type": "object"
      },
...
      "HouseholdResponse": {
        "properties": {
          "baseCurrency": {
            "type": "string"
          },
          "defaultVisibility": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "memberCount": {
            "format": "int32",
            "type": "integer"
          },
          "myMemberId": {
            "format": "uuid",
            "type": "string"
          },
          "myRole": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "baseCurrency",
          "defaultVisibility",
          "id",
          "memberCount",
          "myRole",
          "name",
          "version"
        ],
        "type": "object"
      },
      "InstitutionRow": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isCustom": {
            "type": "boolean"
          },
          "kind": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "isCustom",
          "kind",
          "name"
        ],
        "type": "object"
      },
...
      "LoginResponse": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresInSeconds": {
            "format": "int64",
            "type": "integer"
          },
          "isNewUser": {
            "type": "boolean"
          },
          "refreshToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/MeResponse"
          }
        },
        "required": [
          "accessToken",
          "expiresInSeconds",
          "isNewUser",
          "refreshToken",
          "tokenType",
          "user"
        ],
        "type": "object"
      },
      "MeResponse": {
        "properties": {
          "currency": {
            "type": "string"
          },
          "defaultVisibility": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "currency",
          "defaultVisibility",
          "id",
          "locale"
        ],
        "type": "object"
      },
...
      "OtpChallengeResponse": {
        "properties": {
          "developmentCode": {
            "type": "string"
          },
          "expiresInSeconds": {
            "format": "int64",
            "type": "integer"
          },
          "requestId": {
            "type": "string"
          },
          "resendAfterSeconds": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "expiresInSeconds",
          "requestId",
          "resendAfterSeconds"
        ],
        "type": "object"
      },
      "OtpRequestBody": {
        "properties": {
          "phone": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "phone"
        ],
        "type": "object"
      },
      "OtpVerifyBody": {
        "properties": {
          "code": {
            "pattern": "^[0-9]{4,8}$",
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          },
          "phone": {
            "minLength": 1,
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "phone"
        ],
        "type": "object"
      },
...
      "RefreshBody": {
        "properties": {
          "refreshToken": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "type": "object"
      },
...
      "StepUpStatusResponse": {
        "properties": {
          "elevated": {
            "type": "boolean"
          },
          "expiresInSeconds": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "elevated",
          "expiresInSeconds"
        ],
        "type": "object"
      },
      "StepUpVerifyBody": {
        "properties": {
          "code": {
            "pattern": "^[0-9]{4,8}$",
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      },
      "TaxonomyResponse": {
        "properties": {
          "categoryCode": {
            "type": "string"
          },
          "categoryLabel": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "types": {
            "items": {
              "$ref": "#/components/schemas/TypeResponse"
            },
            "type": "array"
          }
        },
        "required": [
          "categoryCode",
          "categoryLabel",
          "color",
          "icon",
          "types"
        ],
        "type": "object"
      },
      "TokenPair": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresInSeconds": {
            "format": "int64",
            "type": "integer"
          },
          "refreshToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "expiresInSeconds",
          "refreshToken",
          "tokenType"
        ],
        "type": "object"
      },
...
      "TypeResponse": {
        "properties": {
          "categoryCode": {
            "type": "string"
          },
          "categoryLabel": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isCustom": {
            "type": "boolean"
          },
          "label": {
            "type": "string"
          },
          "schema": {
            "$ref": "#/components/schemas/TypeSchema"
          },
          "schemaVersion": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "categoryCode",
          "categoryLabel",
          "code",
          "color",
          "id",
          "isCustom",
          "label",
          "schema",
          "schemaVersion"
        ],
        "type": "object"
      },
...
      "UpdateHouseholdBody": {
        "properties": {
          "defaultVisibility": {
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Access token from POST /api/v1/auth/otp/verify.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "padmini.koduri@bhrigu.tech",
      "name": "Almira"
    },
    "description": "Every rupee a family has invested or owes, in one private place.\n\n**Authentication.** Phone OTP issues a short-lived access token and a\nrotating refresh token. Send `Authorization: Bearer <accessToken>`.\nRefresh tokens are single-use: presenting one twice is treated as theft\nand ends the session.\n\n**Privacy.** Every read passes a per-record visibility check enforced by\nPostgreSQL row-level security. A record you may not see returns 404, not\n403 — a 403 would confirm it exists. Totals are computed through the same\nfilter, so a private record contributes nothing, not even its amount, to\nanother member's figures.\n\n**Money** is decimal, never floating point. Amounts also arrive formatted\n(`₹1,76,875`) so every surface renders them identically.\n\n**Concurrency.** Writes carry the record's `version`; a stale write is\nrejected with 409 and the current version. Creates accept a client-supplied\n`id`, so an offline capture keeps its identity and a retry is idempotent.\n\n**Stability.** v1 is frozen and additive-only. Breaking changes go to v2.",
    "title": "Almira API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/otp/request": {
      "post": {
        "operationId": "requestOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OtpChallengeResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth-controller"
        ]
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "operationId": "verifyOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth-controller"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth-controller"
        ]
      }
    },
...
    "/api/v1/auth/sessions/{id}": {
      "delete": {
        "operationId": "revokeSession",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "auth-controller"
        ]
      }
    },
...
    "/api/v1/auth/step-up/verify": {
      "post": {
        "operationId": "verifyStepUp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepUpVerifyBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StepUpStatusResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth-controller"
        ]
      }
    },
...
    "/api/v1/households/{householdId}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "household-controller"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHouseholdBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "household-controller"
        ]
      }
    },
...
    "/api/v1/households/{householdId}/dashboard": {
      "get": {
        "operationId": "dashboard",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "household",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "member",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "dashboard-controller"
        ]
      }
    },
...
    "/api/v1/households/{householdId}/institutions": {
      "get": {
        "operationId": "institutions",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InstitutionRow"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "catalog-controller"
        ]
      },
      "post": {
        "operationId": "createInstitution",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstitutionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InstitutionRow"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "catalog-controller"
        ]
      }
    },
...
    "/api/v1/households/{householdId}/taxonomy": {
      "get": {
        "operationId": "taxonomy",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaxonomyResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "catalog-controller"
        ]
      }
    },
...
    "/api/v1/households/{householdId}/types": {
      "post": {
        "operationId": "createType",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTypeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TypeResponse"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "catalog-controller"
        ]
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "me",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth-controller"
        ]
      },
      "patch": {
        "operationId": "updateMe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth-controller"
        ]
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ]
}